Refactor and simplify visualizer
